+++ src/cache/commands/CacheTableCommand.js
@@ -18,7 +18,7 @@ class CacheTableCommand extends GeneratorCommand {
 	 * @inheritdoc
 	 */
 	get policies() {
-		return ['db', 'env:local'];
+		return (super.policies || []).concat(['db']);
 	}
 
 	/**