+++ src/console/GeneratorCommand.js
@@ -21,6 +21,13 @@ class GeneratorCommand extends Command {
 	/**
 	 * @inheritdoc
 	 */
+	get policies() {
+		return ['env:local'];
+	}
+
+	/**
+	 * @inheritdoc
+	 */
 	get description() {
 		return `Create a ${this.type} class.`;
 	}
@@ -55,7 +62,11 @@ class GeneratorCommand extends Command {
 	 * @type {string}
 	 */
 	get fileName() {
-		return `${this.parameter('class')}.js`;
+		if (this.parameterIsSupported('class')) {
+			return `${this.parameter('class')}.js`;
+		}
+
+		throw new Error('The "class" parameter does not exists. You must manually define "fileName" accessor.');
 	}
 
 	/**