+++ src/foundation/exceptions/Handler/drivers/OuchDriver.js
@@ -7,6 +7,9 @@ import Driver from './Driver';
 
 /**
  * Ouch exception render driver.
+ *
+ * @memberof foundation.exceptions.Handler.drivers
+ * @augments foundation.exceptions.Handler.drivers.Driver
  */
 class OuchDriver extends Driver {
 
@@ -76,7 +79,7 @@ class OuchDriver extends Driver {
 			data = this.renderJson(exception);
 		}
 
-		if (this.app.environment === 'production') {
+		if (!this.app.isBound('config') || !this.app.make('config').get('app.debug', false)) {
 			const { type, message } = data;
 			data = { type, message };
 		}