+++ src/foundation/exceptions/Handler/drivers/ViewDriver.js
@@ -7,6 +7,9 @@ import Driver from './Driver';
 
 /**
  * Pretty Error exception render driver.
+ *
+ * @memberof foundation.exceptions.Handler.drivers
+ * @augments foundation.exceptions.Handler.drivers.Driver
  */
 class ViewDriver extends Driver {
 
@@ -40,7 +43,7 @@ class ViewDriver extends Driver {
 	renderView(exception) {
 		if (this.app.isBound('view') && this.app.isBound('view.resolver')) {
 			const viewResolver = this.app.make('view.resolver');
-			let viewName = `errors.${exception.status}`;
+			let viewName = `errors.${exception.status || 500}`;
 
 			if (!viewResolver.exists(viewName)) {
 				viewName = 'errors.generic';