+++ src/log/commands/LogTableCommand.js
@@ -18,7 +18,7 @@ class LogTableCommand extends GeneratorCommand {
 	 * @inheritdoc
 	 */
 	get policies() {
-		return ['db', 'env:local'];
+		return (super.policies || []).concat(['db']);
 	}
 
 	/**