+++ src/test/commands/MakeTestCommand.js
@@ -26,13 +26,6 @@ class MakeTestCommand extends GeneratorCommand {
 	/**
 	 * @inheritdoc
 	 */
-	get policies() {
-		return ['env:local'];
-	}
-
-	/**
-	 * @inheritdoc
-	 */
 	get name() {
 		return 'make:test';
 	}